requirements.txt
Django==3.1.3
django-cors-headers==3.5.0
django-phonenumber-field==5.0.0
django-templated-mail==1.1.1
djangorestframework==3.12.2